fix: commits where fail number calculation
this contains the update to the application logic that uses the new CAs we
defined in the previous commit

we're also improving the filtering of skipped tests by using information from
the CAs which allows us to avoid doing more expensive queries during every API
call

we also use window function to calculate slow tests instead of subqueries

I also choose to refactor the code out into modules for each type of data
and a utils module for code reuse
