+++ apps/codecov-api/graphql_api/tests/snapshots/analytics_timescale__TestAnalyticsTestCaseNew__gql_query__0.json
@@ -1,46 +1,6 @@
 [
   {
-    "computed_name": "name0",
-    "testsuite": "testsuite0",
-    "total_pass_count": 1,
-    "total_fail_count": 0,
-    "total_flaky_fail_count": 0,
-    "total_skip_count": 0,
-    "commits_where_fail": 0,
-    "total_count": 1.0,
-    "failure_rate": 0.0,
-    "flake_rate": 0.0,
-    "total_duration": 0.0,
-    "avg_duration": 0.0,
-    "last_duration": 0.0,
-    "flags": [
-      "flag1",
-      "flag2"
-    ],
-    "name": "name0"
-  },
-  {
-    "computed_name": "name1",
-    "testsuite": "testsuite1",
-    "total_pass_count": 0,
-    "total_fail_count": 1,
-    "total_flaky_fail_count": 0,
-    "total_skip_count": 0,
-    "commits_where_fail": 1,
-    "total_count": 1.0,
-    "failure_rate": 1.0,
-    "flake_rate": 0.0,
-    "total_duration": 1.0,
-    "avg_duration": 1.0,
-    "last_duration": 1.0,
-    "flags": [
-      "flag3"
-    ],
-    "name": "name1"
-  },
-  {
-    "computed_name": "name2",
-    "testsuite": "testsuite2",
+    "test_id": "0d312f0fe886f107f94fc93991e2dc87",
     "total_pass_count": 1,
     "total_fail_count": 0,
     "total_flaky_fail_count": 0,
@@ -52,54 +12,17 @@
     "total_duration": 2.0,
     "avg_duration": 2.0,
     "last_duration": 2.0,
+    "last_outcome": "pass",
     "flags": [
       "flag1",
       "flag2"
     ],
-    "name": "name2"
+    "computed_name": "name2",
+    "name": "name2",
+    "testsuite": "testsuite2"
   },
   {
-    "computed_name": "name3",
-    "testsuite": "testsuite3",
-    "total_pass_count": 0,
-    "total_fail_count": 1,
-    "total_flaky_fail_count": 0,
-    "total_skip_count": 0,
-    "commits_where_fail": 1,
-    "total_count": 1.0,
-    "failure_rate": 1.0,
-    "flake_rate": 0.0,
-    "total_duration": 3.0,
-    "avg_duration": 3.0,
-    "last_duration": 3.0,
-    "flags": [
-      "flag3"
-    ],
-    "name": "name3"
-  },
-  {
-    "computed_name": "name4",
-    "testsuite": "testsuite4",
-    "total_pass_count": 1,
-    "total_fail_count": 0,
-    "total_flaky_fail_count": 0,
-    "total_skip_count": 0,
-    "commits_where_fail": 0,
-    "total_count": 1.0,
-    "failure_rate": 0.0,
-    "flake_rate": 0.0,
-    "total_duration": 4.0,
-    "avg_duration": 4.0,
-    "last_duration": 4.0,
-    "flags": [
-      "flag1",
-      "flag2"
-    ],
-    "name": "name4"
-  },
-  {
-    "computed_name": "name5",
-    "testsuite": "testsuite5",
+    "test_id": "16828e614f0b139492d77d5c70d7ecc4",
     "total_pass_count": 0,
     "total_fail_count": 0,
     "total_flaky_fail_count": 0,
@@ -111,7 +34,96 @@
     "total_duration": 0.0,
     "avg_duration": 0.0,
     "last_duration": 0.0,
+    "last_outcome": "skip",
     "flags": null,
-    "name": "name5"
+    "computed_name": "name5",
+    "name": "name5",
+    "testsuite": "testsuite5"
+  },
+  {
+    "test_id": "3158ae556dccb76c50517a257d96886b",
+    "total_pass_count": 1,
+    "total_fail_count": 0,
+    "total_flaky_fail_count": 0,
+    "total_skip_count": 0,
+    "commits_where_fail": 0,
+    "total_count": 1.0,
+    "failure_rate": 0.0,
+    "flake_rate": 0.0,
+    "total_duration": 4.0,
+    "avg_duration": 4.0,
+    "last_duration": 4.0,
+    "last_outcome": "pass",
+    "flags": [
+      "flag1",
+      "flag2"
+    ],
+    "computed_name": "name4",
+    "name": "name4",
+    "testsuite": "testsuite4"
+  },
+  {
+    "test_id": "8817e9110bb1682219b3815ebf02430d",
+    "total_pass_count": 1,
+    "total_fail_count": 0,
+    "total_flaky_fail_count": 0,
+    "total_skip_count": 0,
+    "commits_where_fail": 0,
+    "total_count": 1.0,
+    "failure_rate": 0.0,
+    "flake_rate": 0.0,
+    "total_duration": 0.0,
+    "avg_duration": 0.0,
+    "last_duration": 0.0,
+    "last_outcome": "pass",
+    "flags": [
+      "flag1",
+      "flag2"
+    ],
+    "computed_name": "name0",
+    "name": "name0",
+    "testsuite": "testsuite0"
+  },
+  {
+    "test_id": "dce514d4d3d650ef6c1c34b68af89ec9",
+    "total_pass_count": 0,
+    "total_fail_count": 1,
+    "total_flaky_fail_count": 0,
+    "total_skip_count": 0,
+    "commits_where_fail": 1,
+    "total_count": 1.0,
+    "failure_rate": 1.0,
+    "flake_rate": 0.0,
+    "total_duration": 3.0,
+    "avg_duration": 3.0,
+    "last_duration": 3.0,
+    "last_outcome": "failure",
+    "flags": [
+      "flag3"
+    ],
+    "computed_name": "name3",
+    "name": "name3",
+    "testsuite": "testsuite3"
+  },
+  {
+    "test_id": "ff20823eb233fba7e1b4ad1524fcb87e",
+    "total_pass_count": 0,
+    "total_fail_count": 1,
+    "total_flaky_fail_count": 0,
+    "total_skip_count": 0,
+    "commits_where_fail": 1,
+    "total_count": 1.0,
+    "failure_rate": 1.0,
+    "flake_rate": 0.0,
+    "total_duration": 1.0,
+    "avg_duration": 1.0,
+    "last_duration": 1.0,
+    "last_outcome": "failure",
+    "flags": [
+      "flag3"
+    ],
+    "computed_name": "name1",
+    "name": "name1",
+    "testsuite": "testsuite1"
   }
 ]
